backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.2.0
drf-spectacular==0.26.5
gunicorn==20.1.0
requests==2.31.0
redis==4.6.0
celery==5.2.7
django-redis==5.3.0
flower==1.2.0
django-celery-beat==2.5.0
django-celery-results==2.3.0
reportlab==4.0.7

# Development dependencies
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
django-stubs==4.2.7
pre-commit==3.6.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.4